Add create manby NFT collections method
