+++ src/core/utils/string_utils.rs
@@ -5,9 +5,18 @@ pub fn suffix_str(n: usize, s: impl AsRef<str>) -> String {
     s.as_ref()[split_pos..].to_string()
 }
 
+/// Capitalizes the first character in s.
+pub fn capitalize(s: impl AsRef<str>) -> String {
+    let mut c = s.as_ref().chars();
+    match c.next() {
+        None => String::new(),
+        Some(f) => format!("{}{}", f.to_uppercase(), c.as_str()),
+    }
+}
+
 #[cfg(test)]
 mod tests {
-    use crate::suffix_str;
+    use super::*;
 
     #[test]
     fn str_suffix() {
@@ -17,5 +26,12 @@ mod tests {
     #[test]
     fn string_suffix() {
         assert_eq!(suffix_str(7, "By the rivers of Babylon"), "Babylon")
+    }
+
+    #[test]
+    fn test_capitalize() {
+        assert_eq!(capitalize(""), "");
+        assert_eq!(capitalize("g"), "G");
+        assert_eq!(capitalize("good"), "Good");
     }
 }
