+++ src/wrapped_radix_engine_toolkit/high_level/mod.rs
@@ -2,14 +2,14 @@
 mod assert_manifest;
 
 mod manifests;
-mod manifests_crate_tokens;
+mod manifests_create_tokens;
 mod ret_api;
 mod sargon_specific_types;
 mod token_definition_metadata;
 
 pub use assert_manifest::*;
 pub use manifests::*;
-pub use manifests_crate_tokens::*;
+pub use manifests_create_tokens::*;
 pub use ret_api::*;
 pub use sargon_specific_types::*;
 pub use token_definition_metadata::*;
